+++ promptshell/requirements.txt
@@ -8,3 +8,4 @@ groq
 pyperclip
 pyreadline3
 questionary
+shlex
